Added callback functionality to GameMan.
I think the program should be designed "around" the Game manager, since
that's supposed to be the core of the application.

I think that the peripheral classes should all keep a reference to the
game manager and register callbacks. Let's stick to that design.

Also added test case that confirmed the docs of the STL.