Encapsulate main() logic into MainWindow
I think this is more managable since the various callbacks all need to
access the individual widgets so they need to be in the same scope with
the widget variables. Putting them all in main() will make the function
really tedious.

Also added a primitive "new game" option in the menu.