Implement the UserInputEngine
I realized that there is a hiearchy between the threads that reflects
the ownership relationship between the subsystems. MainWindow owns the
game manager and the game manager owns the engines. By putting the
"threaded entities" at the end of each class, we are able to make sure
that the engine thread terminates before the manager and the manager be-
fore the GUI thread.